requirements.txt
anytree==2.8.0
-e git+git@github.com:sander102907/cpp_ast_parser.git@563988c1bee3f95eacb9edf9bd25f0ab360a63c0#egg=CPP_AST_Parser
gensim==4.0.1
nltk==3.6.2
numpy==1.20.2
pymongo==3.11.4
python-Levenshtein==0.12.2
pytorch-tree-lstm==0.1.3
sacred==0.8.2
sacredboard==0.4.2
sctokenizer==0.0.5
